requirements.txt
langchain
langchain-community
langchain-openai
pydantic
jinja2
sqlalchemy
python-dotenv
